Caesar

Description

This is an implementation of the Caesar cipher.

Documentation

encrypt

Arguments

:: String
The text to encrypt. Note: We check that all chars in the text are in the alphabet. Otherwise we abort the encryption. TODO: Use PlainText as the type.

-> Int
The key to use to encrypt the text. Can be any Int. Pos and Neg. We take the abs of it, before we pass it on.

-> String
Return the encrypted text. TODO: Use EncryptedText as the type.

Encrypt a given text with a given key.

decrypt

Arguments

:: String
The text to encrypt.

-> Int
The key to use.

-> String
Return the encrypted text.

Decrypt a given text with a given key.

alphabet :: String

The alphabet (means all of the chars) we support for encryption. |TODO: Use type Alphabet as the type (instead of String).